Conflicts:
	CMakeLists.txt
	source/settings/Settings.h

+++ source/tracking/imagePreProcessor/ImagePreProcessor.cpp
@@ -17,7 +17,7 @@ void ImagePreProcessor::init()
 {
     _pMOG = std::make_unique<cv::BackgroundSubtractorMOG2>(
         _settings.getValueOfParam<int>(PREPROCESSORPARAM::BG_MOG2_HISTORY),
-        _settings.getValueOfParam<int>(PREPROCESSORPARAM::BG_MOG2_VAR_THRESHOLD),
+        _settings.getValueOfParam<float>(PREPROCESSORPARAM::BG_MOG2_VAR_THRESHOLD),
         _settings.getValueOfParam<bool>(PREPROCESSORPARAM::BG_MOG2_SHADOW_DETECTION));
 }
